requirements.txt
requests
beautifulsoup4
cloudscraper
python-telegram-bot


